requirements.txt
fastapi
uvicorn
pypdf
sentence-transformers
faiss-cpu
python multi-part